move data and remove resistType from data
